260320 GET /archive/{videoId} 신설. 이것저것 DTO추가 등등 자잘자잘한 수정.

+++ src/main/java/com/haconaka/demo/repository/member/MemberRepo.java
@@ -3,6 +3,6 @@ package com.haconaka.demo.repository.member;
 import com.haconaka.demo.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepo extends JpaRepository<MemberEntity, Integer> {
+public interface MemberRepo extends JpaRepository<MemberEntity, Long> {
     MemberEntity findByYoutubeChannelId(String channelId);
 }
